move initialize.go to provider/
