+++ pkg/app/pipedv1/plugin/terraform/provider/initialize_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package deployment
+package provider
 
 import (
 	"testing"
